+++ example/hello-templating/webapp/classic.rsp.dart
@@ -15,7 +15,7 @@ Future classic(HttpConnect connect, {header, sidebar, body, footer}) { //#2
   <div class="header">
     """); //#2
 
-  response.write(RSP.nns(header)); //#4
+  response.write(RSP.nnx(header)); //#4
 
 
   response.write("""
@@ -24,7 +24,7 @@ Future classic(HttpConnect connect, {header, sidebar, body, footer}) { //#2
   <div class="sidebar">
     """); //#4
 
-  response.write(RSP.nns(sidebar)); //#7
+  response.write(RSP.nnx(sidebar)); //#7
 
 
   response.write("""
@@ -33,7 +33,7 @@ Future classic(HttpConnect connect, {header, sidebar, body, footer}) { //#2
   <div class="body">
     """); //#7
 
-  response.write(RSP.nns(body)); //#10
+  response.write(RSP.nnx(body)); //#10
 
 
   response.write("""
@@ -42,7 +42,7 @@ Future classic(HttpConnect connect, {header, sidebar, body, footer}) { //#2
   <div class="footer">
     """); //#10
 
-  response.write(RSP.nns(footer)); //#13
+  response.write(RSP.nnx(footer)); //#13
 
 
   response.write("""