+++ test/features/webapp/searchResult.rsp.dart
@@ -22,28 +22,28 @@ Future searchResult(HttpConnect connect, {criteria}) { //#2
     <ul>
       <li>text: """); //#2
 
-  response.write(RSP.nns(criteria.text)); //#11
+  response.write(RSP.nnx(criteria.text)); //#11
 
 
   response.write("""
 </li>
       <li>since: """); //#11
 
-  response.write(RSP.nns(criteria.since)); //#12
+  response.write(RSP.nnx(criteria.since)); //#12
 
 
   response.write("""
 </li>
       <li>within: """); //#12
 
-  response.write(RSP.nns(criteria.within)); //#13
+  response.write(RSP.nnx(criteria.within)); //#13
 
 
   response.write("""
 </li>
       <li>hasAttachment: """); //#13
 
-  response.write(RSP.nns(criteria.hasAttachment)); //#14
+  response.write(RSP.nnx(criteria.hasAttachment)); //#14
 
 
   response.write("""