+++ test/syntax/lastModified1.rsp.dart
@@ -14,7 +14,7 @@ Future lastModified1(HttpConnect connect, {input}) { //#2
 
   if (!connect.isIncluded)
     response.headers.contentType = ContentType.parse("""text/html; charset=utf-8""");
-  response.headers.set(HttpHeaders.LAST_MODIFIED, new DateTime.fromMillisecondsSinceEpoch(1370408104117));
+  response.headers.set(HttpHeaders.LAST_MODIFIED, new DateTime.fromMillisecondsSinceEpoch(1371204582266));
 
   response.write("""
 <html>